Base implementation of spawners
Spawners now function, they spawn the wave enemylist with a 1 second delay between each enemy.
Spawner.cs has a OnDrawGizmos function to make it easier to see where they are in the scene view.
Time between spawns can be changed for finetuning.

Spawner has been renamed to WaveManager as thats a better description of its role.
Uses a index that increments with each wave to get the next set of enemies to spawn.
StartNextWave is a public method (can be changed to private if the condition checking is going to be handled by the WaveManager itself.

Level.cs default value for strength multiplier changed to 1.1f

GameManager now holds the level data and has a wave field for the ui to access.