Implementation of slot scripts
Trying to get the lists to work together, and the slots to be updated on changes but im doing it a bit wierdly, should actually switch over to events.
currently the problem is that on IMGUI press of Update Slots and adding of a item in the inventory list gives a null reference error.

UIManager and ItemPickup have had minor updates on QoL (auto checking and name consistincy)