Working Item Drops
Item's are now dropped by enemies, allowing the player to up the stats of their tower.

- Chance of drop is determined by a ratio (which is the distance to tower divided by spawner distance to tower) multiplied by a base drop chance (curerntly 50% for debugging purposes)
- Light and Material colour change depending on rarity

ItemTable ScriptableObject that holds lists for all items and their different rarities. Has a method for getting a random item in that rarity. Allows for adding elements in the future.
Only need one material and itemprefab now due to instantiation of a material for each itempickup.